Updated openstack/openstack Project: openstack/nova  6f8928c779fe3ac16d9100122cf4cbc927ab0e48
libvirt: add comment for vifs_already_plugged=True in finish_migration

Given we unplug VIFs in _cleanup_resize if the host has changed, it's
confusing to pass vifs_already_plugged=True in finish_migration without
some explanation of why we don't check to see if the host has changed
before deciding what the value of vifs_already_plugged should be when
creating the new domain and plugging VIFs again.

After some discussion with Neutron cores, this is the explanation I've
come up with, albeit not a great one, and indicates a gap in some of the
nova/neutron integration story.

Related-Bug: #1323658

Change-Id: I982b6bd0650f29f34c213b6eebbe066f7bd65815