Updated openstack/openstack Project: openstack/nova  e23d0dfa7f6226aa23939e166bac25aca604d4ae
Keep instance state if lvm backend not impl

Migration is not implemented for LVM backed instances
on libvirt driver, so precheck will raise exception
but that will make the instance enter ERROR state.
Actually, the VM itself is not changed, the resize
action failed at preparation stage, so don't need to
set it to 'ERROR' state.

Change-Id: I3274972c969a014b4f090c5e60ec1ebcb35962a0