Updated openstack/openstack Project: openstack/keystone  65ad9a85c5c5cae7715a054bba67976399e88e6f
Add unscoped token formatter for Fernet tokens

In the initial Fernet token implementation, scoped and unscoped tokens were
grouped under the same schema. This change improves the distinction between
unscoped token schemas and scoped token schemas. This makes the logic for each
formatter simpler since we don't have determine if the token has a project
in it or not. The formatter should only have to deal with the schema that's
passed to it.

partially implements klw-tokens

Change-Id: I0eef686cee4665b2e03892cbc1a80d7e7d6be9d9