Updated openstack/openstack Project: openstack/tempest  d9ce2b523a0a5c0ab3b3171cc56da946ede00c86
Switch swift tests to use cred_provider with roles

The swift tests were using a DataGenerator class to have users with
special roles for performing certain swift operations. This class never
really took into account the cred_provider mechanism and just brute
forced things with admin. This commit address this by removing the
swift data generator and changing all it's uses to get_creds_by_roles()
where needed.

As part of this cleanup many of the tests were also relying on an
implicit assignment of the operator role with tenant isolation. As part
of making role requirements explicit in this patch that had to be
removed.

Change-Id: I07fa785ef0bad9677885488f3155df7d47f8c846
Partially-implements: bp test-accounts-continued