Updated openstack/openstack Project: openstack-dev/devstack  40ce320bb013f850a47d32781dd2f77a4d7927fa
Temporarily stop building mumpy wheel

It turns out we aren't actually using this wheel since we are still
installing the deb python-numpy, and building numpy takes several
minutes which is a lot considering we do it on every single dsvm job.

So until we have wheel caching in place, and are actually using the
version we build ourselves, stop wasting time.

Change-Id: I7643c55598e5ecc29ea708c537818b37a8047d4b