Updated openstack/openstack Project: openstack/horizon  19c93f92eab38b5ff78a29acac65c953366157ad
Migrate all instances from host marked for maintenance

This patch adds migrate capability to Horizon for host
already marked for maintenance.

All instances could be cold migrated.
There is an option also for running instance to allow
making live migration to them.
All running instances will be migrated as the same
migrating configuration, if the administrator wants to
migrate a specific instance in a specific configuration he
could do it from the instance dashboard.

blueprint migrate-all-instances-from-hosts-in-maintenance-mode

Change-Id: Ia1260831e79ede66a9d4320b092bebeb023796bc
Co-Authored-By: Bartosz Fic <bartosz.fic@intel.com>