Updated openstack/openstack Project: openstack/cinder  610c672998c87e2bf2f919a4fdea86749c748a1f
HP 3par driver filter and evaluator function

This patch implements the merged change in the scheduler
https://review.openstack.org/#/c/129987/ in HP 3par driver.

1. Added a way to get filter_function and goodness_function strings from
cinder.conf in the base class VolumeDriver,  for example

filter_function =
"capabilities.total_volumes < 600 && capabilities.capacity_utilization < 80"
goodness_function = "capabilities.capacity_utilization < 40 ? 80 : 60"

The strings can be got from the [DEFAULT] section or a driver instance group
section.

If the string can not be found in the cinder.conf, it will try to find a
default filter_fuction or a default goodness_function which can be specified
in the driver code. HP3par drivers don't provide default values. However if
other driver want to provide default values, it can be done by overwriting
get_default_filter_function and get_default_goodness_function.

2. Also added goodness_function and filter_function to the
_update_volume_stats in ISCSIDriver, FibreChannelDriver, FakeISCSIDriver
and LVMVolumeDriver.

3. Added 2 stats total_volumes and capacity_utilization so they can be
used as the part of the formula in filter_function or goodness_function
string.

4. Fixed free_capacity for limited cpg to include SDUsage's usedMiB as part of
the calcuation so it is consistent with unlimited cpg and capacity_utilization
looks better.

5. Added total_volumes capability to the LVM driver.

Implements: blueprint hp3par-driver-supplies-filtering-weighing-functions
Change-Id: I4ce77c9c1b1e14de82465bb94068b2ff10c19b91