Updated openstack/openstack Project: openstack/openstackdocstheme  0b19390f6691580f44147cbdafe2322fc8261716
Fix for unwanted plus sign in sublists

A plus sign with gray background appears next to a bulleted list inside
a numbered list. It does NOT appear in all other sublist cases such as
bullet list nested inside numbered list.

The plus sign is caused by CSS class "docs-has-sub", defined in
combined.css.

Example - given this RST

    #. Number 1
    #. Number 2

       * Bullet 1

the following HTML (or similar) is rendered:

    <ol>
    <li> Number 1
    <li class="docs-has-sub">Number 2
       <ul>
       <li>Bullet 1
    ....

Where does the "docs-has-sub" come from? It has been added to the <li>
element by the following line in docs.js:

    $('ol > li:has(ul)').addClass('docs-has-sub');

The purpose of this line seems to be implementation of an expandable and
collapsable sublist. The plus sign indicates to the user that the list
can be expanded. Perhaps this implementation is incomplete?

Since the Javascript line affects all bulleted sublists of numbered
lists, not only expandable ones, I am removing it. The expansion feature
has to be implemented differently.

Closes-Bug: #1417291
Change-Id: Ib6ecd176ca0e1c788a9277ebbf38ea91b6b891ce