Updated openstack/openstack Project: openstack/horizon  ae1fc154c57dd233f8a6b3b95d0378d737260380
edit option hided when firewall in pending status

if the firewall in pending status, it is not allowed
to update. It is throwing an error. so the horizon
should not allow to update or edit the firewall
during the pending status.

this patch hiding the edit firewall option, if the
firewall in pending status.

Change-Id: I4ffb0cf2bd135e815bbaf24bed13b49cbeb36019
closes-Bug: #1427594