Updated openstack/openstack Project: openstack/keystone  c4f13f37d12856a1857feb60a65375dca2da36f2
Reload drivers when their domain config is updated

When a domain config is modifed, in order for those configs
to actually be live, we need to re-initialise the backend
identity driver for the respective domain.

This patch uses cache timeouts to ensure that new config options
become visible to all keystone processes/threads.

Partially Implements: blueprint domain-config-ext
Change-Id: I1a802dde53f25f945eadd235574e9c342a770443