Updated openstack/openstack Project: openstack-infra/project-config  69cf58216c4364af731270146c20945964d6edb2
Ironic: Move pxe_ipa-nv to check instead of experimental

The pxe_ipa-nv job is testing the future default ramdisk for Ironic, at
first we added it to experimetal just to make sure it works. Now that
we know it's passing we are moving it to the "check" so it can run on
every gate check for us to collect data and make it voting in the future.

Change-Id: I7187bdb1480cafa2f8152d0b0229a436c1fa3514