Updated openstack/openstack Project: openstack/ossa  cf981cbfc5fb4e44375c8d3b58b01a3d38da6a75
Add extra info on landing page

Reduce the list of advisories on landing page to five most recent
advisories, and add extra info taken from the website landing page:
http://www.openstack.org/projects/openstack-security/

Move the complete OSSA list to a separate page. This is the first step
of the work moving out information from the security page on the main
website to security.o.o.

Change-Id: Iabaab1891988c71aaf1bcfb1390adb4c2ad8d531