Updated openstack/openstack Project: openstack/keystone  085144a0df4bbd850462d3b5c5a9f48e8bcbb1f3
Deprecate passing "extras" in token data

The "extras" attribute in v3 was introduced to copy a bug in v2 which
leaked an implementation detail of the SQL backend (the "extras" column)
into token responses. I can't find any evidence that this has ever been
useful (or used) in v3.

bp: deprecated-as-of-kilo
Change-Id: I4c1162b39d0f5c29069ec06921a69cddb70c041b