Updated openstack/openstack Project: openstack-infra/nodepool  9af49ae8a35c0dd655b255376e7331234edeba18
Don't allocate nodes to providers without images

To build nodes for labels providers need to have a current image for
that label available. If the provider does not have a current image do
not add it to the allocation requests for this label.

This requires some changes to how we wait for resources in the tests. We
can no longer assume an image will become available then be allocated
to nodes and wait for those nodes. Instead we have to explicitly wait
for the images then wait for the nodes.

Change-Id: I4802c06c25b4019a6159bac993b6d7d4934fe6ba