Updated openstack/openstack Project: openstack/neutron  4663a15efd83a64274ebe3c9b9f9e852732d57ca
Raise error upon deleting subnet with router ports

Fixes an issue where SLAAC and DHCPV6-stateless subnets can be
deleted even if they are attached to an internal router port.
This patch raises an exception whenever a subnet is deleted
that has existing IP Allocations on an internal router port.

Change-Id: I0a16156274b5736236654fca6700ef2d67f4519b
Closes-Bug: #1424760