Updated openstack/openstack Project: openstack/horizon  97603552b3fb66ae86ab477e8a007441160b6593
Always show stack status reason in topology

Currently, only when the stack is in error status, the stack status
reason will be displayed, otherwise, there will be two line with
exactly same words, and this makes the enduser confuse. So fixes
this by changing to not show anything when stack in not in error
status.

Change-Id: Ifcd8def6b452ecccef9a8daa0e31504cffa2d6ad
Closes-bug: #1428580