Updated openstack/openstack Project: openstack-infra/project-config  692be9e322578a0806a8c6e96279addae6951ee9
Print instance UUID to log

This macro allows a job running on a cloud-based instance to print the
UUID of that instance to the build log.  By the time it is determined
that a test failed due to a system or networking error, the instance
has already been deleted.  The UUID can be used to locate the physical
system the instance ran on so that analysis can be done at that level.
By providing it in the build log, users will not have to find someone
with access to the nodepool logs to get this level of support for this
type of issue.

Depends-On: Ib84ec30b5c4c5de5d5be80cae2b4230c1aba3bc6
Change-Id: Iddcd56e2e2f2a942573a9868d1c009106b38564a