Updated openstack/openstack Project: openstack/oslo.config  dc28858f353531444a4f65ffd4330615dd26892a
print better message when choices has an empty string

Follow on to:
I3bb3621420915a6a2990440294a389b45a23c519

choices can have None (as a default) or an empty string "" (as a
default value or when read from a config file). We need to print
better message in the sample config for these two cases.

Change-Id: Iba483c0e4180fc95181b161f996cce5e627035f5