Updated openstack/openstack Project: openstack/python-openstackclient  2f4cfb6cd50e6305111fdf51698fb73835b9aa58
Move OSC auth plugins so they can be found

The plugin detection at the top of openstackclient.api.auth did not
detect the plugins at the bottom of that file because, surprise,
they had not been declared yet so the entry points were ignored.

Move both plugin subclasses into openstackclient.api.auth_plugin.

Fix a problem with the password callback that was otherwise made
worse with this change.

Closes-Bug: 1428912
Change-Id: Idc3b72534071e0013c8922884a8bc14137509a0f