Updated openstack/openstack Project: openstack/neutron  f9313697548a06ad1b10297f49d81754d52d8749
In Arista ML2 delete tenant without any resources

If there are no resources associated with a
tenant, it was deleted from Arista DB in
delete_xxx_precommit() methods. The sync
mechanism used to make sure that back-end is in
sync with DB.

Lately, the sync mechanism has been enhanced to
support scaled deployment. Therefore, the
delete_tenat() call now needs to be invoked from
delete_xxx_postcommit() methods instead of
delete_xxx_precommit() methods. Additionally,
when the tenant is removed from the DB, it is
removed from the back-end as well. This removes
the dependancy on the sync mechanism to take
care of this.
[Note: Sync will still do it, but, this way it
gets done right away]

Closes-bug: 1429968
Change-Id: I2f6ddb01079fe4e2648685ca97d0be06db2d1a55