Updated openstack/openstack Project: openstack/horizon  d2426dfceb4046ec7a713fe7dd36640d251435f5
Refactor multiple references to $(document)

According to the Horizon Contributing Guide, we are to:
"Avoid creating instances of the same object repeatedly
within the same scope. Instead, assign the object to a
variable and re-use the existing object."

The guide references the use of "$(document)", as an
example. And yet multiple examples can be found of that
sort of approach. This commit cleans up those references.

Partial-Bug: 1412971
Change-Id: Ica05700101ae9a453623d928f07b75fa7aaca186