Updated openstack/openstack Project: openstack-dev/devstack  b8c74c11083a6b2ada9f0a3fafd1c67e1d5d845b
Remove setting use_floating_ips values

We remove the code to set use_floating_ips. In old code it was set
false if we do not use neutron. However, we cannot deploy clusters
with floating ips by that. So we just use the default value, which
is set True in Sahara.

Closes-Bug: #1426226
Change-Id: Idfcdc5ab776681ddc740dc12035e04da349ea089