Updated openstack/openstack Project: openstack/ceilometer  df10a87ea7810f3ebf47b7d027e30a8403d89b0f
alarming: add gnocchi alarm rules

Testing the API part of Gnocchi alarms within the Gnocchi tree is
a pain and doesn't really have senses.

So, this change moves the gnocchi alarms rules and evaluator that live
in Gnocchi tree to Ceilometer tree.

This also permits to add these new alarm rules into the documentation.

And add some tests that cannot be done on the gnocchi side.

DocImpact

Change-Id: I1bbc9f904d55b51cd1b3c51dcdfaf58f01bd9075