Updated openstack/openstack Project: openstack/python-barbicanclient  f48efbdf5302c59b401f64ca5c46665a3093bb6a
Do not filter empty strings

Empty strings shouldn't be filtered from requests, as we have
explicitly set some attributes to be optional and actually
differentiate between null and empty attributes in the server side. So
now, only attributes containing None are filtered and empty strings
are actually passed to the server.

Change-Id: Ia048d7de6bcbc12317930eeccc009b6b0db10e13
Closes-bug: #1420445