Updated openstack/openstack Project: openstack/keystone  d3916f7787bc87cda5e355b0e6d5686851cf867f
Fix seconds since epoch use in fernet tokens

The fernet code wasn't using seconds since the epoch correctly.
The seconds since epoch doesn't depend on the timezone, but the
time int calculated by the fernet functions was based on the local
time of the system.

This was causing local unit tests to fail.

Closes-Bug: 1429663
Change-Id: I56757e9636e7baf46eeb1657dab209616e310672