Updated openstack/openstack Project: openstack/python-barbicanclient  8437e47288816219ace14235ac04d011b1d782fa
Fix delete a secret that does not exist

Changes the behavior of deleting a secret that does not exist.
The secret uuid is now verified as a valid UUID. If it not a valid
uuid it will result in a value error. A valid uuid will be allowed
to go through to the server and a 404 will be returned if the secret
does not exist.

Change-Id: I35e7fa36714604c2beba044d0e8bb492233bb7a4