Updated openstack/openstack Project: openstack/python-novaclient  b7972478ff656bc6fa1cc909b617b02b0dfb28b7
Update help messages for default security group commands

The secgroup-{add,delete,list}-default-rules? family of commands has
confusing help text that makes it sounds like they actually operate on
the default security group (that is, the security group named
'default' which is the default for the current tenant). Instead, the
operate on a sort of meta-security-group object which is only used to
populate the 'default' security group in newly-created tenants. This
is crazy confusing. These commands get fancy new help strings that
should hopefully make it a little more clear.

Closes-Bug: 1430354
Change-Id: I4d9569ba7ea9ad7a20e984df7a137c99240a9748