Updated openstack/openstack Project: openstack/horizon  8ad24206f993c7c8a76600545d86ecc7b8cfb84c
fixed row update for loadbalancer tables

In loadbalancer when creating pool or member
the corresponding row is not updating as per
the status change

manual page reload is needed to see the
current status

this patch added a row update class.
this will update the current row as per
the status change.

Change-Id: I4a3fed098e2e04aec5d72d83618a40e086b0a16c
Closes-Bug: #1425057