Updated openstack/openstack Project: openstack/horizon  8e00e11a5ab3cda8e19f7b11e5e0638dd754e612
Narrow selector to eliminate JavaScript keydown perf hit

The users data (/identity/users/) can get quite large in
some environments. In one of our environments, we have 800+
users. The users table on the index page is not paginated,
and so shows all 800+ users.

If you attempt to create a new user using the "Create User"
modal form, keyboard performance dies. It can take several
seconds for each keystroke to register in each input field
on the form.

It turns out that there is are several event handlers bound
to the "keydown" JavaScript event, at the level of the
document itself. This has been done to ensure that dynamic
content gets handled properly. In the case where we have
several hundred rows of data on the page, this is killing
performance.

The solution: bind the event handler to the highest level
element in the document necessary to handle the required
event. The focus of the JavaScript in play here is to
enable keyboard navigation between tabs in a form. Tab
groups will always have a DIV with a class attribute of
"tab-content". So we now bind our event handlers to that
class instead of the overall document.

Closes-Bug: #1411892
Change-Id: I9a37fb20a2506b1dabffc5f2b0b7b97341de5335