Updated openstack/openstack Project: openstack/ceilometer  287377a41db939820d26c4eac2e486721f9fe143
normalise timestamp in query

when type casting query parameters, we use parse_isotime. this has
a side effect of adding a tzinfo to the resulting datetime. no db
currently stores tzinfo so we should normalise query to drop this
information. as a side effect, it also adds in missing supported type

Change-Id: Id4a5ffbb5388122b43adcf5054a7cd0cce3d275d
Closes-Bug: #1423545